Optimize Code
To speed up the use of the dashboard and its processes, certain changes
were made:

Pandas and Numpy vectorization were used for the download data and
process_data subroutines.

Groups are now only flagged as removed instead of removed from the
dataset when deleting invalid groups from the dashboard. These are then removed from
the dataset when uploading the groups to the stored CSV on the cloud.